add support for Redshift API; code refactoring
