+++ localstack/mock/infra.py
@@ -10,16 +10,15 @@ import logging
 import requests
 import json
 import boto3
-import random
-import urlparse
-from requests.models import Response
 import __init__
+from localstack import constants
 from localstack.utils.aws import aws_stack
 from localstack.utils import common
 from localstack.utils.common import *
-from localstack.mock import firehose_api, lambda_api, generic_proxy, dynamodbstreams_api
+from localstack.mock import generic_proxy
+from localstack.mock.apis import firehose_api, lambda_api, dynamodbstreams_api
+from localstack.mock.proxy import apigateway_listener, dynamodb_listener, kinesis_listener, sns_listener
 from localstack.mock.generic_proxy import GenericProxy
-from localstack import constants
 
 THIS_PATH = os.path.dirname(os.path.realpath(__file__))
 ROOT_PATH = os.path.realpath(os.path.join(THIS_PATH, '..'))
@@ -28,17 +27,6 @@ ROOT_PATH = os.path.realpath(os.path.join(THIS_PATH, '..'))
 SIGNAL_HANDLERS_SETUP = False
 INFRA_STOPPED = False
 
-# cache table definitions - used for testing
-TABLE_DEFINITIONS = {}
-
-# mappings for SNS topic subscriptions
-# TODO possibly move to a separate file?
-SNS_SUBSCRIPTIONS = {}
-
-# constants
-KINESIS_ACTION_PUT_RECORD = 'Kinesis_20131202.PutRecord'
-KINESIS_ACTION_PUT_RECORDS = 'Kinesis_20131202.PutRecords'
-
 INSTALL_DIR_INFRA = '%s/infra' % ROOT_PATH
 INSTALL_DIR_NPM = '%s/node_modules' % ROOT_PATH
 INSTALL_DIR_ES = '%s/elasticsearch' % INSTALL_DIR_INFRA
@@ -46,7 +34,7 @@ TMP_ARCHIVE_ES = '/tmp/localstack.es.zip'
 
 # list of default APIs to be spun up
 DEFAULT_APIS = ['s3', 'sns', 'sqs', 'es', 'apigateway', 'dynamodb',
-    'kinesis', 'dynamodbstreams', 'firehose', 'lambda']
+    'kinesis', 'dynamodbstreams', 'firehose', 'lambda', 'redshift']
 
 # set up logger
 LOGGER = logging.getLogger(__name__)
@@ -157,28 +145,34 @@ def start_elasticsearch(port=DEFAULT_PORT_ELASTICSEARCH, delete_data=True, async
 
 def start_apigateway(port=DEFAULT_PORT_APIGATEWAY, async=False, update_listener=None):
     backend_port = DEFAULT_PORT_APIGATEWAY_BACKEND
-    cmd = '%s/bin/moto_server apigateway -p%s' % (LOCALSTACK_VENV_FOLDER, backend_port)
+    cmd = '%s/bin/moto_server apigateway -p %s -H %s' % (LOCALSTACK_VENV_FOLDER, backend_port, constants.BIND_HOST)
     print("Starting mock API Gateway (port %s)..." % port)
     start_proxy(port, backend_port, update_listener)
     return do_run(cmd, async)
 
 
 def start_s3(port=DEFAULT_PORT_S3, async=False):
-    cmd = '%s/bin/moto_server s3 -p%s' % (LOCALSTACK_VENV_FOLDER, port)
+    cmd = '%s/bin/moto_server s3 -p %s -H %s' % (LOCALSTACK_VENV_FOLDER, port, constants.BIND_HOST)
     print("Starting mock S3 server (port %s)..." % port)
+    return do_run(cmd, async)
+
+
+def start_redshift(port=DEFAULT_PORT_REDSHIFT, async=False):
+    cmd = '%s/bin/moto_server redshift -p %s -H %s' % (LOCALSTACK_VENV_FOLDER, port, constants.BIND_HOST)
+    print("Starting mock Redshift server (port %s)..." % port)
     return do_run(cmd, async)
 
 
 def start_sns(port=DEFAULT_PORT_SNS, async=False, update_listener=None):
     backend_port = DEFAULT_PORT_SNS_BACKEND
-    cmd = '%s/bin/moto_server sns -p%s' % (LOCALSTACK_VENV_FOLDER, backend_port)
+    cmd = '%s/bin/moto_server sns -p %s -H %s' % (LOCALSTACK_VENV_FOLDER, backend_port, constants.BIND_HOST)
     print("Starting mock SNS server (port %s)..." % port)
     start_proxy(port, backend_port, update_listener)
     return do_run(cmd, async)
 
 
 def start_sqs(port=DEFAULT_PORT_SQS, async=False):
-    cmd = '%s/bin/moto_server sqs -p%s' % (LOCALSTACK_VENV_FOLDER, port)
+    cmd = '%s/bin/moto_server sqs -p %s -H %s' % (LOCALSTACK_VENV_FOLDER, port, constants.BIND_HOST)
     print("Starting mock SQS server (port %s)..." % port)
     return do_run(cmd, async)
 
@@ -324,13 +318,13 @@ def start_infra(async=False, dynamodb_update_listener=None, kinesis_update_liste
         apigateway_update_listener=None, sns_update_listener=None, apis=DEFAULT_APIS):
     try:
         if not dynamodb_update_listener:
-            dynamodb_update_listener = update_dynamodb
+            dynamodb_update_listener = dynamodb_listener.update_dynamodb
         if not kinesis_update_listener:
-            kinesis_update_listener = update_kinesis
+            kinesis_update_listener = kinesis_listener.update_kinesis
         if not apigateway_update_listener:
-            apigateway_update_listener = update_apigateway
+            apigateway_update_listener = apigateway_listener.update_apigateway
         if not sns_update_listener:
-            sns_update_listener = update_sns
+            sns_update_listener = sns_listener.update_sns
         # set environment
         os.environ['AWS_REGION'] = DEFAULT_REGION
         os.environ['ENV'] = ENV_DEV
@@ -364,6 +358,8 @@ def start_infra(async=False, dynamodb_update_listener=None, kinesis_update_liste
             thread = start_lambda(async=True)
         if 'kinesis' in apis:
             thread = start_kinesalite(async=True, update_listener=kinesis_update_listener)
+        if 'redshift' in apis:
+            start_redshift(async=True)
         # Elasticsearch and S3 take a bit to come up
         time.sleep(3)
         # check that all infra components are up and running
@@ -382,192 +378,6 @@ def start_infra(async=False, dynamodb_update_listener=None, kinesis_update_liste
     finally:
         if not async:
             stop_infra()
-
-
-def update_apigateway(method, path, data, headers, response=None, return_forward_info=False):
-    if return_forward_info:
-        regex1 = r'^/restapis/[A-Za-z0-9\-]+/deployments$'
-        if method == 'POST' and re.match(regex1, path):
-            # this is a request to deploy the API gateway, simply return HTTP code 200
-            return 200
-
-        regex2 = r'^/restapis/([A-Za-z0-9_\-]+)/([A-Za-z0-9_\-]+)/%s/([^/]+)$' % PATH_USER_REQUEST
-        if method == 'POST' and re.match(regex2, path):
-            api_id = re.search(regex2, path).group(1)
-            sub_path = '/%s' % re.search(regex2, path).group(3)
-            integration = aws_stack.get_apigateway_integration(api_id, method, sub_path)
-            template = integration['requestTemplates'][APPLICATION_JSON]
-            new_request = aws_stack.render_velocity_template(template, data)
-
-            # forward records to our main kinesis stream
-            # TODO check whether the target of this API method is 'kinesis'
-            headers = aws_stack.mock_aws_request_headers(service='kinesis')
-            headers['X-Amz-Target'] = KINESIS_ACTION_PUT_RECORDS
-            result = common.make_http_request(url=TEST_KINESIS_URL,
-                method='POST', data=new_request, headers=headers)
-            return 200
-        return True
-
-
-def update_sns(method, path, data, headers, response=None, return_forward_info=False):
-    if return_forward_info:
-        if method == 'POST' and path == '/':
-            req_data = urlparse.parse_qs(data)
-            topic_arn = req_data.get('TopicArn')
-            if topic_arn:
-                topic_arn = topic_arn[0]
-                if topic_arn not in SNS_SUBSCRIPTIONS:
-                    SNS_SUBSCRIPTIONS[topic_arn] = []
-            if 'Subscribe' in req_data['Action']:
-                subscription = {
-                    'topic_arn': topic_arn,
-                    'endpoint': req_data['Endpoint'][0],
-                    'protocol': req_data['Protocol'][0]
-                }
-                SNS_SUBSCRIPTIONS[topic_arn].append(subscription)
-            elif 'Publish' in req_data['Action']:
-                message = req_data['Message'][0]
-                sqs_client = aws_stack.connect_to_service('sqs', client=True)
-                for subscriber in SNS_SUBSCRIPTIONS[topic_arn]:
-                    if subscriber['protocol'] == 'sqs':
-                        queue_name = subscriber['endpoint'].split(':')[5]
-                        queue_url = subscriber.get('sqs_queue_url')
-                        if not queue_url:
-                            queue_url = aws_stack.get_sqs_queue_url(queue_name)
-                            subscriber['sqs_queue_url'] = queue_url
-                        sqs_client.send_message(QueueUrl=queue_url, MessageBody=message)
-                    else:
-                        LOGGER.warning('Unexpected protocol "%s" for SNS subscription' % subscriber['protocol'])
-                # return response here because we do not want the request to be forwarded to SNS
-                response = Response()
-                response._content = """<PublishResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">
-                    <PublishResult><MessageId>n/a</MessageId></PublishResult>
-                    <ResponseMetadata><RequestId>n/a</RequestId></ResponseMetadata></PublishResponse>"""
-                response.status_code = 200
-                return response
-        return True
-
-
-def kinesis_error_response(data):
-    error_response = Response()
-    error_response.status_code = 200
-    content = {"FailedRecordCount": 1, "Records": []}
-    for record in data["Records"]:
-        content["Records"].append({
-            "ErrorCode": "ProvisionedThroughputExceededException",
-            "ErrorMessage": "Rate exceeded for shard X in stream Y under account Z."
-        })
-    error_response._content = json.dumps(content)
-    return error_response
-
-
-def update_kinesis(method, path, data, headers, response=None, return_forward_info=False):
-    if return_forward_info:
-        if random.random() < constants.KINESIS_ERROR_PROBABILITY:
-            return kinesis_error_response(data)
-        else:
-            return True
-
-    action = headers['X-Amz-Target'] if 'X-Amz-Target' in headers else None
-    records = []
-    if action == KINESIS_ACTION_PUT_RECORD:
-        record = {
-            'data': data['Data'],
-            'partitionKey': data['PartitionKey']
-        }
-        records = [record]
-        stream_name = data['StreamName']
-        lambda_api.process_kinesis_records(records, stream_name)
-    elif action == KINESIS_ACTION_PUT_RECORDS:
-        records = []
-        for record in data['Records']:
-            record = {
-                'data': record['Data'],
-                'partitionKey': record['PartitionKey']
-            }
-            records.append(record)
-        stream_name = data['StreamName']
-        lambda_api.process_kinesis_records(records, stream_name)
-
-
-def update_dynamodb(method, path, data, headers, response=None, return_forward_info=False):
-    if return_forward_info:
-        return True
-
-    # update table definitions
-    if data and 'TableName' in data and 'KeySchema' in data:
-        TABLE_DEFINITIONS[data['TableName']] = data
-
-    action = headers.get('X-Amz-Target')
-    if not action:
-        return
-
-    response_data = json.loads(response.text)
-    record = {
-        "eventID": "1",
-        "eventVersion": "1.0",
-        "dynamodb": {
-            "StreamViewType": "NEW_AND_OLD_IMAGES",
-            "SequenceNumber": "1",
-            "SizeBytes": -1
-        },
-        "awsRegion": DEFAULT_REGION,
-        "eventSource": "aws:dynamodb"
-    }
-    event = {
-        'Records': [record]
-    }
-
-    if action == 'DynamoDB_20120810.UpdateItem':
-        req = {'TableName': data['TableName']}
-        req['Key'] = data['Key']
-        new_item = aws_stack.dynamodb_get_item_raw(TEST_DYNAMODB_URL, req)
-        if 'Item' not in new_item:
-            if 'message' in new_item:
-                print('WARNING: Unable to get item from DynamoDB: %s' % new_item['message'])
-            return
-        record['eventName'] = 'MODIFY'
-        record['dynamodb']['Keys'] = data['Key']
-        record['dynamodb']['NewImage'] = new_item['Item']
-    elif action == 'DynamoDB_20120810.PutItem':
-        record['eventName'] = 'INSERT'
-        keys = dynamodb_extract_keys(item=data['Item'], table_name=data['TableName'])
-        record['dynamodb']['Keys'] = keys
-        record['dynamodb']['NewImage'] = data['Item']
-    elif action == 'DynamoDB_20120810.DeleteItem':
-        record['eventName'] = 'REMOVE'
-        record['dynamodb']['Keys'] = data['Key']
-    elif action == 'DynamoDB_20120810.CreateTable':
-        if 'StreamSpecification' in data:
-            stream = data['StreamSpecification']
-            enabled = stream['StreamEnabled']
-            if enabled:
-                table_name = data['TableName']
-                view_type = stream['StreamViewType']
-                dynamodbstreams_api.add_dynamodb_stream(table_name=table_name,
-                    view_type=view_type, enabled=enabled)
-        return
-    else:
-        # nothing to do
-        return
-    record['eventSourceARN'] = aws_stack.dynamodb_table_arn(data['TableName'])
-    sources = lambda_api.get_event_sources(source_arn=record['eventSourceARN'])
-    if len(sources) > 0:
-        pass
-    for src in sources:
-        func_to_call = lambda_api.lambda_arn_to_function[src['FunctionArn']]
-        lambda_api.run_lambda(func_to_call, event=event, context={})
-
-
-def dynamodb_extract_keys(item, table_name):
-    result = {}
-    if table_name not in TABLE_DEFINITIONS:
-        print("WARN: unknown table: %s not found in %s" % (table_name, TABLE_DEFINITIONS))
-        return None
-    for key in TABLE_DEFINITIONS[table_name]['KeySchema']:
-        attr_name = key['AttributeName']
-        result[attr_name] = item[attr_name]
-    return result
 
 
 if __name__ == '__main__':
