+++ requirements.txt
@@ -7,7 +7,7 @@ elasticsearch==2.3.0
 flask==0.10.1
 flask_swagger==0.2.12
 jsonpath-rw==1.4.0
-moto-ext==0.4.26
+moto-ext==0.4.32
 nose==1.3.7
 pep8==1.7.0
 python-coveralls==2.7.0
