+++ setup.py
@@ -76,7 +76,7 @@ if __name__ == '__main__':
 
     setup(
         name='localstack',
-        version='0.3.1',
+        version='0.3.2',
         description='An easy-to-use test/mocking framework for developing Cloud applications',
         author='Waldemar Hummer (Atlassian)',
         author_email='waldemar.hummer@gmail.com',
